pyproject.toml
[tool.poetry]
name = "telegram-video-circle-bot"
version = "0.1.0"
description = ""
authors = ["Slava Kuryaev <92830706+necrosskull@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "telegram_video_circle_bot"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = "20.6"
moviepy = "^1.0.3"
opencv-python = "^4.8.1.78"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
